refactor(sandbox-repro): probe imports shipped resolvePresetRoot + narrow Windows CI to canary role
Two changes based on review of the CI's purpose:

1. Extract `resolvePresetRoot` into its own file (`server/plugins/
   resolvePresetRoot.ts`) so the sandbox probe can import the SHIPPED
   resolver instead of an inline copy. Breaking the fix in production
   now breaks the probe — the whole point of Windows CI. `preset-loader.ts`
   re-exports it so existing unit test / server callers are unaffected.

2. Convert probe from `.mjs` to `.ts` and import the shipped resolver
   directly. `resolvePresetRoot.ts`'s dep footprint is deliberately
   tiny (`node:fs` / `node:module` / `node:path`) so a node:22-slim
   container with a global `tsx` install can load it without the full
   server graph (logger, plugin registry). `test/tsconfig.json`
   excludes `sandbox-repro/` because the probe's import path
   (`/app/server-plugins/resolvePresetRoot.ts`) only exists inside
   the container.

3. Narrow the workflow triggers to `schedule` + `workflow_dispatch`
   only. Every PR to preset-loader or the sandbox mount config is
   already covered by the POSIX unit test
   (`test/plugins/test_preset_loader_node_path.ts`) on the
   ubuntu/macos matrix. Windows CI's incremental value is catching
   upstream environment drift (yarn switching from junctions to
   hardlinks, WSL2 changing how junctions surface, Docker changing
   bind-mount semantics) — slow-moving, so a daily canary suffices.

Docs updated to reflect the two-tier coverage (POSIX unit for code
correctness, Windows CI for env drift) and the probe's new architecture.
